feat: support diffusion dpo
