+++ align_anything/configs/train/text+image_to_text/rm.yaml
