+++ align_anything/configs/train/text_to_text/kto.yaml
