+++ align_anything/datasets/text_image_to_text/prompt_only.py
@@ -20,38 +20,51 @@ from typing_extensions import TypedDict  # Python 3.10+
 
 import torch
 import transformers
-from torchvision import transforms
 from torch.utils.data import Dataset
 from transformers.tokenization_utils import PaddingStrategy, TruncationStrategy
-
+from torchvision import transforms
 from align_anything.utils.multi_process import get_current_device
 from align_anything.utils.template_registry import get_template_class
-from align_anything.utils.tools import right_padding
+from align_anything.utils.tools import left_padding
 from datasets import load_dataset
 
 
 IGNORE_INDEX = -100
 
 __all__ = [
-    'SupervisedDataset',
-    'SupervisedCollator',
-    'SupervisedSample',
-    'SupervisedBatch',
+    'PromptOnlyDataset',
+    'PromptOnlyCollator',
+    'PromptOnlySample',
+    'PromptOnlyBatch',
 ]
 
 
-class SupervisedSample(TypedDict, total=True):
+def remove_duplicate_prompts(dict_list: list[dict[str, Any]], template):
+    seen_prompts = set()
+    unique_dict_list = []
+    for idx in range(len(dict_list)):
+        item = dict_list[idx]
+        prompt = template.format_prompt_only_sample(item)['text']
+        if prompt not in seen_prompts:
+            unique_dict_list.append(item)
+            seen_prompts.add(prompt)
+    return unique_dict_list
+
+
+class PromptOnlySample(TypedDict, total=True):
     input_ids: torch.LongTensor  # size = (L,)
     labels: torch.LongTensor  # size = (L,)
+    pixel_values: torch.LongTensor | None  # size = (B, C, H, W)
 
 
-class SupervisedBatch(TypedDict, total=True):
+class PromptOnlyBatch(TypedDict, total=True):
     input_ids: torch.LongTensor  # size = (B, L)
     labels: torch.LongTensor  # size = (B, L)
     attention_mask: torch.BoolTensor  # size = (B, L)
+    pixel_values: torch.LongTensor | None  # size = (B, C, H, W)
 
 
-class SupervisedDataset(Dataset):
+class PromptOnlyDataset(Dataset):
 
     def __init__(
         self,
@@ -63,20 +76,21 @@ class SupervisedDataset(Dataset):
         split: str | None = None,
         subset: str | None = None,
         data_files: str | None = None,
-        optional_args: list | str = [],
     ):
         super().__init__()
         assert path, f'You must set the valid datasets path! Here is {path}'
         assert template, f'You must set the valid template path! Here is {template}'
         self.tokenizer = tokenizer
         self.processor = processor
-        self.raw_data = load_dataset(path, split=split, data_files=data_files, subset=subset,  *optional_args, trust_remote_code=True)
-        if size:
-            self.raw_data = self.raw_data.select(range(int(size)))
+        raw_data_duplicated = load_dataset(path, split=split, data_files=data_files, subset=subset,  *optional_args, trust_remote_code=True)
         self.template = get_template_class(template)
+        self.raw_data = remove_duplicate_prompts(raw_data_duplicated, self.template)
 
-    def preprocess(self, raw_sample: dict[str, Any]) -> SupervisedSample:
-        formatted_sample = self.template.format_sample(raw_sample)
+        if size:
+            self.raw_data = self.raw_data[:size]
+
+    def preprocess(self, raw_sample: dict[str, Any]) -> PromptOnlySample:
+        formatted_sample = self.template.format_prompt_only_sample(raw_sample)
         return_dict = {}
         raw_text = ''
         if isinstance(formatted_sample['text'], list):
@@ -87,16 +101,15 @@ class SupervisedDataset(Dataset):
             raise NotImplementedError
         return_dict['input_ids'] = self.tokenize(raw_text)
 
-        formatted_prompt = formatted_sample['prompt']
-        labels = return_dict['input_ids'].clone()
-        # mask non-assistant input
-        labels[: len(self.tokenize(formatted_prompt))] = IGNORE_INDEX
-        return_dict['labels'] = labels
+        raw_image = formatted_sample['image']
+        return_dict['pixel_values'] = self.processor.image_processor(
+            raw_image, return_tensors='pt'
+        )['pixel_values'][0]
 
         return return_dict
 
     def get_collator(self) -> Callable[[list[dict[str, torch.Tensor]]], dict[str, torch.Tensor]]:
-        return SupervisedCollator(self.tokenizer.pad_token_id)
+        return PromptOnlyCollator(self.tokenizer.pad_token_id)
 
     def tokenize(
         self,
@@ -122,7 +135,7 @@ class SupervisedDataset(Dataset):
     def __getitem__(self, index: int) -> dict[str, torch.Tensor]:
         """Get a tokenized data sample by index."""
         raw_sample = self.raw_data[index]
-        data = self.preprocess(raw_sample.copy())
+        data = self.preprocess(raw_sample)
         return data
 
     def __len__(self) -> int:
@@ -130,28 +143,30 @@ class SupervisedDataset(Dataset):
         return len(self.raw_data)
 
 
-class SupervisedCollator:
+class PromptOnlyCollator:
 
     def __init__(self, pad_token_id: int) -> None:
         """Initialize a collator."""
         self.pad_token_id = pad_token_id
 
-    def __call__(self, samples: list[SupervisedSample]) -> SupervisedBatch:
+    def __call__(self, samples: list[PromptOnlySample]) -> PromptOnlyBatch:
         return_dict = {}
         current_device = get_current_device()
 
-        return_dict['input_ids'] = right_padding(
-            [sample['input_ids'] for sample in samples],
-            padding_value=self.pad_token_id,
-        ).to(current_device)
+        input_ids = [sample['input_ids'] for sample in samples]
+        attention_mask = [
+            input_id.new_ones(input_id.size(), dtype=torch.bool) for input_id in input_ids
+        ]
 
-        return_dict['labels'] = right_padding(
-            [sample['labels'] for sample in samples],
-            padding_value=IGNORE_INDEX,
-        ).to(current_device)
-
-        return_dict['attention_mask'] = (
-            return_dict['input_ids'].ne(self.pad_token_id).to(current_device)
+        return_dict['input_ids'] = left_padding(input_ids, padding_value=self.pad_token_id).to(
+            current_device
         )
+        return_dict['attention_mask'] = left_padding(attention_mask, padding_value=0).to(
+            current_device
+        )
+
+        return_dict['pixel_values'] = torch.stack(
+            [sample['pixel_values'] for sample in samples]
+        ).to(current_device)
 
         return return_dict
