+++ align_anything/datasets/text_to_image/preference.py
@@ -20,8 +20,8 @@ from typing_extensions import TypedDict  # Python 3.10+
 
 import torch
 import transformers
-from torchvision import transforms
 from torch.utils.data import Dataset
+from torchvision import transforms
 from transformers.tokenization_utils import PaddingStrategy, TruncationStrategy
 
 from align_anything.utils.multi_process import get_current_device
@@ -29,36 +29,38 @@ from align_anything.utils.template_registry import get_template_class
 from align_anything.utils.tools import right_padding
 from datasets import load_dataset
 
+from PIL import Image
 
-IGNORE_INDEX = -100
 
 __all__ = [
-    'SupervisedDataset',
-    'SupervisedCollator',
-    'SupervisedSample',
-    'SupervisedBatch',
+    'PreferenceDataset',
+    'PreferenceCollator',
+    'PreferenceSample',
+    'PreferenceBatch',
 ]
 
 
-class SupervisedSample(TypedDict, total=True):
+class PreferenceSample(TypedDict, total=True):
     input_ids: torch.LongTensor  # size = (L,)
     labels: torch.LongTensor  # size = (L,)
+    pixel_values: torch.LongTensor | None  # size = (B, C, H, W)
 
 
-class SupervisedBatch(TypedDict, total=True):
+class PreferenceBatch(TypedDict, total=True):
     input_ids: torch.LongTensor  # size = (B, L)
     labels: torch.LongTensor  # size = (B, L)
     attention_mask: torch.BoolTensor  # size = (B, L)
+    pixel_values: torch.LongTensor | None  # size = (B, C, H, W)
 
 
-class SupervisedDataset(Dataset):
+class PreferenceDataset(Dataset):
 
     def __init__(
         self,
         path: str,
         template: str,
         tokenizer: transformers.PreTrainedTokenizer,
-        processor: transformers.ProcessorMixin | transforms.Compose | None = None,
+        processor: transforms.Compose | None = None,
         size: int | None = None,
         split: str | None = None,
         subset: str | None = None,
@@ -69,34 +71,43 @@ class SupervisedDataset(Dataset):
         assert path, f'You must set the valid datasets path! Here is {path}'
         assert template, f'You must set the valid template path! Here is {template}'
         self.tokenizer = tokenizer
-        self.processor = processor
-        self.raw_data = load_dataset(path, split=split, data_files=data_files, subset=subset,  *optional_args, trust_remote_code=True)
-        if size:
-            self.raw_data = self.raw_data.select(range(int(size)))
         self.template = get_template_class(template)
-
-    def preprocess(self, raw_sample: dict[str, Any]) -> SupervisedSample:
+        self.transforms = processor
+        
+        if isinstance(optional_args, str):
+            optional_args = [optional_args]
+        self.raw_data = load_dataset(path, split=split, data_files=data_files, subset=subset,  *optional_args, trust_remote_code=True)
+        self.valid_indices = self.fillter_indices()
+        
+        if size:
+            size = min(size, len(self.raw_data))
+            self.raw_data = self.raw_data.select(range(int(size)))
+        
+    def fillter_indices(self):
+        valid_indices = []
+        for i, item in enumerate(self.raw_data):
+            if not self.template.check_equal(item):
+                valid_indices.append(i)
+        return valid_indices
+        
+    def preprocess(self, raw_sample: dict[str, Any]) -> PreferenceSample:
         formatted_sample = self.template.format_sample(raw_sample)
         return_dict = {}
-        raw_text = ''
-        if isinstance(formatted_sample['text'], list):
-            raw_text = self.tokenizer.eos_token.join(formatted_sample['text'])
-        elif isinstance(formatted_sample['text'], str):
-            raw_text = formatted_sample['text'] + self.tokenizer.eos_token
-        else:
-            raise NotImplementedError
-        return_dict['input_ids'] = self.tokenize(raw_text)
-
-        formatted_prompt = formatted_sample['prompt']
-        labels = return_dict['input_ids'].clone()
-        # mask non-assistant input
-        labels[: len(self.tokenize(formatted_prompt))] = IGNORE_INDEX
-        return_dict['labels'] = labels
-
+        
+        return_dict['input_ids'] = self.tokenize(formatted_sample['prompt'], add_special_tokens=False)
+        better_pixel_values = self.process_image(formatted_sample['better_image'])
+        worse_pixel_values = self.process_image(formatted_sample['worse_image'])
+        
+        all_pixel_values = torch.cat([better_pixel_values, worse_pixel_values], dim=0)
+        return_dict['pixel_values'] = all_pixel_values
+        
         return return_dict
 
+    def process_image(self, raw_image: Image) -> torch.Tensor:
+        return self.transforms(raw_image)
+
     def get_collator(self) -> Callable[[list[dict[str, torch.Tensor]]], dict[str, torch.Tensor]]:
-        return SupervisedCollator(self.tokenizer.pad_token_id)
+        return PreferenceCollator(self.tokenizer.pad_token_id)
 
     def tokenize(
         self,
@@ -109,7 +120,7 @@ class SupervisedDataset(Dataset):
         """Tokenize a text string into a tensor representation."""
         if max_length is None:
             max_length = self.tokenizer.model_max_length
-
+        
         return self.tokenizer(
             text,
             add_special_tokens=add_special_tokens,
@@ -121,22 +132,23 @@ class SupervisedDataset(Dataset):
 
     def __getitem__(self, index: int) -> dict[str, torch.Tensor]:
         """Get a tokenized data sample by index."""
-        raw_sample = self.raw_data[index]
-        data = self.preprocess(raw_sample.copy())
+        valid_idx = self.valid_indices[index]
+        raw_sample = self.raw_data[valid_idx]
+        data = self.preprocess(raw_sample)
         return data
 
     def __len__(self) -> int:
         """Get the number of samples in the dataset."""
-        return len(self.raw_data)
+        return len(self.valid_indices)
 
 
-class SupervisedCollator:
+class PreferenceCollator:
 
     def __init__(self, pad_token_id: int) -> None:
         """Initialize a collator."""
         self.pad_token_id = pad_token_id
 
-    def __call__(self, samples: list[SupervisedSample]) -> SupervisedBatch:
+    def __call__(self, samples: list[PreferenceSample]) -> PreferenceSample:
         return_dict = {}
         current_device = get_current_device()
 
@@ -145,13 +157,8 @@ class SupervisedCollator:
             padding_value=self.pad_token_id,
         ).to(current_device)
 
-        return_dict['labels'] = right_padding(
-            [sample['labels'] for sample in samples],
-            padding_value=IGNORE_INDEX,
-        ).to(current_device)
-
-        return_dict['attention_mask'] = (
-            return_dict['input_ids'].ne(self.pad_token_id).to(current_device)
-        )
-
+        return_dict['pixel_values'] = torch.stack(
+            [sample['pixel_values'] for sample in samples]
+        ).to(current_device).to(memory_format=torch.contiguous_format).float()
+        
         return return_dict
