+++ align_anything/models/pretrained_model.py
@@ -35,14 +35,9 @@ from align_anything.models.model_registry import AnyModel
 from align_anything.utils.multi_process import is_main_process, get_current_device
 
 from accelerate.state import AcceleratorState
-from diffusers import AutoencoderKL, DDPMScheduler, StableDiffusionPipeline, UNet2DConditionModel
-from diffusers.optimization import get_scheduler
-from diffusers.training_utils import EMAModel, compute_dream_and_update_latents, compute_snr
-from diffusers.utils import check_min_version, deprecate, is_wandb_available, make_image_grid
-from diffusers.utils.hub_utils import load_or_create_model_card, populate_model_card
-from diffusers.utils.import_utils import is_xformers_available
-from diffusers.utils.torch_utils import is_compiled_module
+from diffusers import AutoencoderKL, DDPMScheduler, UNet2DConditionModel
 
+from peft import LoraConfig
 from transformers.utils import ContextManagers
 from transformers import CLIPTextModel, CLIPTokenizer
 
@@ -230,8 +225,6 @@ def load_pretrained_models(  # pylint: disable=too-many-arguments
 
 def load_pretrained_diffusion_models(  # pylint: disable=too-many-arguments
     model_name_or_path: str | os.PathLike,
-    model_max_length: int = 512,
-    padding_side: Literal['left', 'right'] = 'right',
     dtype: torch.dtype = torch.bfloat16,
     *,
     cache_dir: str | os.PathLike | None = None,
@@ -240,19 +233,20 @@ def load_pretrained_diffusion_models(  # pylint: disable=too-many-arguments
     non_ema_revision: str | None = None,
     variant: str | None = None,
     freeze_unet: bool = False,
+    lora_unet: bool = False,
+    lora_rank: int = 8,
 ) -> tuple[PreTrainedModel, PreTrainedTokenizerBase]:
     """Load pre-trained model and tokenizer from a given path."""
     model_name_or_path = os.path.expanduser(model_name_or_path)
     # Load scheduler, tokenizer and models.
     noise_scheduler = DDPMScheduler.from_pretrained(model_name_or_path, subfolder="scheduler")
-    tokenizer = CLIPTokenizer.from_pretrained(
+    tokenizer = AutoTokenizer.from_pretrained(
         model_name_or_path, 
         subfolder="tokenizer", 
         revision=revision,
-        model_max_length=model_max_length,
         cache_dir=cache_dir,
-        padding_side=padding_side,
         trust_remote_code=trust_remote_code,
+        use_fast=False,
     )
     def deepspeed_zero_init_disabled_context_manager():
         """
@@ -281,9 +275,25 @@ def load_pretrained_diffusion_models(  # pylint: disable=too-many-arguments
         unet.requires_grad_(False)
     else:
         unet.train()
-    
     current_device = get_current_device()
     text_encoder.to(current_device, dtype=dtype)
     vae.to(current_device, dtype=dtype)
+    
+    if lora_unet:
+        # Set up LoRA.
+        unet_lora_config = LoraConfig(
+            r=lora_rank,
+            lora_alpha=lora_rank,
+            init_lora_weights="gaussian",
+            target_modules=["to_k", "to_q", "to_v", "to_out.0"],
+        )
+        # Add adapter and make sure the trainable params are in float32.
+        unet.add_adapter(unet_lora_config)
+        for param in unet.parameters():
+            # only upcast trainable parameters (LoRA) into fp32
+            if param.requires_grad:
+                param.data = param.to(torch.float32)
+    
+    resize_tokenizer_embedding(tokenizer=tokenizer, model=text_encoder)
     
     return unet, vae, text_encoder, noise_scheduler, tokenizer
