+++ align_anything/trainers/base/__init__.py
@@ -12,9 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Dataset classes for text to text training."""
 
 
-from align_anything.datasets.text_to_text.preference import *
-from align_anything.datasets.text_to_text.prompt_only import *
-from align_anything.datasets.text_to_text.supervised import *
+from align_anything.trainers.base.supervised_trainer import SupervisedTrainerBase
+from align_anything.trainers.base.rl_trainer import RLTrainerBase
