--- align_anything/trainers/image_to_text/dpo.py
@@ -1,430 +1,0 @@
-# Copyright 2024 PKU-Alignment Team. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""Trainer for DPO training."""
-
-
-import argparse
-import os
-import sys
-from datetime import datetime
-from typing import Any
-
-import deepspeed
-import torch
-import torch.distributed as dist
-import torch.nn.functional as F
-from deepspeed.ops.adam import FusedAdam
-from torch.utils.data import DataLoader
-from torch.utils.data.distributed import DistributedSampler
-from tqdm import tqdm
-from transformers import CONFIG_NAME, AutoModelForCausalLM, PreTrainedModel, get_scheduler
-from transformers.integrations.deepspeed import HfDeepSpeedConfig
-
-from align_anything.datasets.text_to_text.preference import PreferenceBatch, PreferenceDataset
-from align_anything.models.pretrained_model import load_pretrained_models
-from align_anything.utils.logger import Logger
-from align_anything.utils.multi_process import (
-    get_all_reduce_mean,
-    get_current_device,
-    is_main_process,
-)
-from align_anything.utils.tools import (
-    custom_cfgs_to_dict,
-    dict_to_namedtuple,
-    gather_log_probabilities,
-    get_optimizer_grouped_parameters,
-    namedtuple_to_dict,
-    prepare_ds_eval_cfgs,
-    prepare_ds_train_cfgs,
-    read_cfgs,
-    seed_everything,
-    update_dict,
-)
-
-
-class DPOTrainer:
-
-    def __init__(self, cfgs, ds_cfgs) -> None:
-        """Initialize trainer."""
-        self.cfgs = cfgs
-        self.ds_train_cfgs = prepare_ds_train_cfgs(custom_cfgs=cfgs.train_cfgs, raw_ds_cfgs=ds_cfgs)
-        self.ds_eval_cfgs = prepare_ds_eval_cfgs(custom_cfgs=cfgs.train_cfgs, raw_ds_cfgs=ds_cfgs)
-        self.global_step = 0
-
-        self.init_check()
-        dist.barrier()
-        self.init_models()
-        dist.barrier()
-        self.init_datasets()
-        dist.barrier()
-        self.init_engines()
-        dist.barrier()
-        self.init_logger()
-
-    def init_check(self) -> None:
-        """Initial configuration checking."""
-        return
-
-    def init_logger(self) -> None:
-        """Set logger."""
-        logger_cfgs = self.cfgs.logger_cfgs
-        time = datetime.now().strftime('%Y-%m-%d-%H-%M-%S')
-
-        self.logger = Logger(
-            log_type=logger_cfgs.log_type,
-            log_dir=logger_cfgs.output_dir,
-            log_project=logger_cfgs.log_project,
-            log_run_name=f'{logger_cfgs.log_run_name}-{self.cfgs.data_cfgs.train_datasets}-{time}',
-            config=namedtuple_to_dict(self.cfgs),
-        )
-
-    def init_models(self) -> None:
-        """Initialize model and tokenizer."""
-        if self.ds_train_cfgs['zero_optimization']['stage'] == 3:
-            self.dstchf_train = HfDeepSpeedConfig(self.ds_train_cfgs)
-        if self.ds_eval_cfgs['zero_optimization']['stage'] == 3:
-            self.dsechf_eval = HfDeepSpeedConfig(self.ds_eval_cfgs)
-        self.model, self.tokenizer, self.processor = load_pretrained_models(
-            self.cfgs.model_cfgs.model_name_or_path,
-            model_max_length=self.cfgs.model_cfgs.model_max_length,
-            padding_side='left',
-            trust_remote_code=self.cfgs.train_cfgs.trust_remote_code,
-            freeze_mm_proj=self.cfgs.train_cfgs.freeze_mm_proj,
-            freeze_vision_tower=self.cfgs.train_cfgs.freeze_vision_tower,
-        )
-        self.reference_model, _, _ = load_pretrained_models(
-            self.cfgs.model_cfgs.model_name_or_path,
-            model_max_length=self.cfgs.model_cfgs.model_max_length,
-            padding_side='left',
-            trust_remote_code=self.cfgs.train_cfgs.trust_remote_code,
-        )
-
-    def init_datasets(self) -> None:
-        """Initialize training and evaluation datasets."""
-        train_dataset = PreferenceDataset(
-            path=self.cfgs.data_cfgs.train_datasets,
-            template=self.cfgs.data_cfgs.train_template,
-            tokenizer=self.tokenizer,
-            processor=self.processor,
-            size=self.cfgs.data_cfgs.train_size,
-            split=self.cfgs.data_cfgs.train_split,
-            subset=self.cfgs.data_cfgs.train_subset,
-            data_files=self.cfgs.data_cfgs.train_data_files,
-        )
-        self.train_dataloader = DataLoader(
-            train_dataset,
-            collate_fn=train_dataset.get_collator(),
-            sampler=DistributedSampler(train_dataset, shuffle=True),
-            batch_size=self.cfgs.train_cfgs.per_device_train_batch_size,
-        )
-        if self.cfgs.data_cfgs.eval_datasets:
-            eval_dataset = PreferenceDataset(
-                path=self.cfgs.data_cfgs.eval_datasets,
-                template=self.cfgs.data_cfgs.eval_template,
-                tokenizer=self.tokenizer,
-                processor=self.processor,
-                size=self.cfgs.data_cfgs.eval_size,
-                split=self.cfgs.data_cfgs.eval_split,
-                subset=self.cfgs.data_cfgs.eval_subset,
-                data_files=self.cfgs.data_cfgs.eval_data_files,
-            )
-            self.eval_dataloader = DataLoader(
-                eval_dataset,
-                collate_fn=eval_dataset.get_collator(),
-                sampler=DistributedSampler(eval_dataset, shuffle=True),
-                batch_size=self.cfgs.train_cfgs.per_device_train_batch_size,
-            )
-        self.split_token = train_dataset.template.split_token
-
-    def init_engines(self) -> None:
-        """Initialize DeepSpeed engines."""
-        num_update_steps_per_epoch = (
-            len(self.train_dataloader) + self.cfgs.train_cfgs.gradient_accumulation_steps - 1
-        ) // self.cfgs.train_cfgs.gradient_accumulation_steps
-        total_training_steps = self.cfgs.train_cfgs.epochs * num_update_steps_per_epoch
-
-        optimizer_grouped_parameters = get_optimizer_grouped_parameters(
-            self.model,
-            self.cfgs.train_cfgs.weight_decay,
-        )
-        optimizer = FusedAdam(
-            optimizer_grouped_parameters,
-            lr=self.cfgs.train_cfgs.learning_rate,
-            betas=self.cfgs.train_cfgs.adam_betas,
-        )
-
-        num_warmup_steps = int(self.cfgs.train_cfgs.lr_warmup_ratio * total_training_steps)
-        lr_scheduler = get_scheduler(
-            name=self.cfgs.train_cfgs.lr_scheduler_type,
-            optimizer=optimizer,
-            num_warmup_steps=num_warmup_steps,
-            num_training_steps=total_training_steps,
-        )
-
-        self.model, *_ = deepspeed.initialize(
-            model=self.model,
-            optimizer=optimizer,
-            config=self.ds_train_cfgs,
-            lr_scheduler=lr_scheduler,
-            dist_init_required=True,
-        )
-
-        self.reference_model, *_ = deepspeed.initialize(
-            model=self.reference_model,
-            config=self.ds_eval_cfgs,
-        )
-
-        if self.cfgs.train_cfgs.gradient_checkpointing:
-            self.model.gradient_checkpointing_enable()
-
-    @staticmethod
-    def compute_log_probs(
-        model: AutoModelForCausalLM,
-        batch: PreferenceBatch,
-    ) -> torch.Tensor:
-        """Compute log probabilities of given sequences."""
-        logits = model(**batch).logits
-        input_ids = batch['input_ids']
-        return gather_log_probabilities(logits[:, :-1], input_ids[:, 1:])
-
-    def loss(  # pylint: disable=too-many-locals
-        self,
-        batch: PreferenceBatch,
-    ) -> dict[str, torch.Tensor]:
-        """Loss function for the DPO algorithm."""
-        sequence_log_probs = self.compute_log_probs(
-            self.model.module,
-            batch,
-        )
-        (
-            better_sequence_log_probs,  # size = (B, L - 1)
-            worse_sequence_log_probs,  # size = (B, L - 1)
-        ) = sequence_log_probs.chunk(chunks=2, dim=0)
-
-        with torch.no_grad():
-            ref_sequence_log_probs = self.compute_log_probs(  # size = (2 * B, L - 1)
-                self.reference_model.module,
-                batch,
-            )
-            ref_better_sequence_log_probs, ref_worse_sequence_log_probs = (
-                ref_sequence_log_probs.chunk(chunks=2, dim=0)
-            )
-
-        losses = []
-        better_sample_rewards = []
-        worse_sample_rewards = []
-
-        better_input_ids, worse_input_ids = batch['input_ids'].chunk(chunks=2, dim=0)
-        better_attention_mask, worse_attention_mask = batch['attention_mask'].chunk(chunks=2, dim=0)
-
-        batch_size = better_input_ids.size(0)
-        for i in range(batch_size):
-            if torch.all(torch.eq(better_input_ids[i], worse_input_ids[i])).item():
-                continue
-            better_end_index = better_attention_mask[i].nonzero()[-1].squeeze().item()
-            worse_end_index = worse_attention_mask[i].nonzero()[-1].squeeze().item()
-            diverge_index = (
-                (better_input_ids[i] != worse_input_ids[i]).nonzero()[0].squeeze().item()
-            )
-            assert 0 <= diverge_index <= better_end_index, 'diverge index is out of range!'
-            assert 0 <= diverge_index <= worse_end_index, 'diverge index is out of range!'
-
-            better_seq_slice = slice(diverge_index, better_end_index + 1)
-            worse_seq_slice = slice(diverge_index, worse_end_index + 1)
-
-            # size = ()
-            better_log_prob = better_sequence_log_probs[i, better_seq_slice].sum(dim=-1)
-            worse_log_prob = worse_sequence_log_probs[i, worse_seq_slice].sum(dim=-1)
-            ref_better_log_prob = ref_better_sequence_log_probs[i, better_seq_slice].sum(dim=-1)
-            ref_worse_log_prob = ref_worse_sequence_log_probs[i, worse_seq_slice].sum(dim=-1)
-            better_log_ratio = better_log_prob - ref_better_log_prob
-            worse_log_ratio = worse_log_prob - ref_worse_log_prob
-
-            losses.append(
-                -F.logsigmoid(
-                    self.cfgs.train_cfgs.scale_coeff * (better_log_ratio - worse_log_ratio),
-                ),
-            )
-            better_sample_rewards.append(
-                self.cfgs.train_cfgs.scale_coeff * better_log_ratio.detach(),
-            )
-            worse_sample_rewards.append(self.cfgs.train_cfgs.scale_coeff * worse_log_ratio.detach())
-
-        loss = torch.stack(losses).mean()  # size = ()
-        better_sample_reward = torch.stack(better_sample_rewards)  # size = (B,)
-        worse_sample_reward = torch.stack(worse_sample_rewards)  # size = (B,)
-        reward = better_sample_reward + worse_sample_reward  # size = (B,)
-        reward_accuracy = (better_sample_reward > worse_sample_reward).float().mean()  # size = ()
-        reward_margin = better_sample_reward - worse_sample_reward  # size = (B,)
-
-        return {
-            'loss': loss,
-            'reward': reward,
-            'better_sample_reward': better_sample_reward,
-            'worse_sample_reward': worse_sample_reward,
-            'reward_accuracy': reward_accuracy,
-            'reward_margin': reward_margin,
-        }
-
-    def train_step(
-        self,
-        batch: PreferenceBatch,
-    ) -> dict[str, Any]:
-        """Perform a single training step for DPO."""
-        loss_dict = self.loss(batch=batch)
-        loss = loss_dict['loss']
-        self.model.backward(loss)
-        self.model.step()
-
-        with torch.no_grad():
-            reward = loss_dict['reward'].mean()
-            better_sample_reward = loss_dict['better_sample_reward'].mean()
-            worse_sample_reward = loss_dict['worse_sample_reward'].mean()
-            reward_accuracy = loss_dict['reward_accuracy']
-            reward_margin = loss_dict['reward_margin'].mean()
-
-            loss = get_all_reduce_mean(loss)
-            reward = get_all_reduce_mean(reward)
-            better_sample_reward = get_all_reduce_mean(better_sample_reward)
-            worse_sample_reward = get_all_reduce_mean(worse_sample_reward)
-            reward_accuracy = get_all_reduce_mean(reward_accuracy)
-            reward_margin = get_all_reduce_mean(reward_margin)
-
-        return {
-            'train/loss': loss.item(),
-            'train/reward': reward.item(),
-            'train/better_sample_reward': better_sample_reward.item(),
-            'train/worse_sample_reward': worse_sample_reward.item(),
-            'train/reward_accuracy': reward_accuracy.item(),
-            'train/reward_margin': reward_margin.item(),
-            'train/lr': self.model.optimizer.param_groups[0]['lr'],
-        }
-
-    @torch.no_grad()
-    def eval(self) -> dict[str, Any]:
-        """Evaluate the model on the evaluation dataset."""
-        return {}
-
-    def train(self) -> None:
-        """Train the model."""
-        self.logger.print('***** Running training *****')
-
-        progress_bar = tqdm(
-            total=self.cfgs.train_cfgs.epochs * len(self.train_dataloader),
-            desc=f'Training 1/{self.cfgs.train_cfgs.epochs} epoch',
-            position=0,
-            leave=True,
-            disable=not is_main_process(),
-        )
-
-        if self.cfgs.data_cfgs.eval_datasets:
-            self.logger.print('\n***** Evaluating at the beginning *****')
-            self.logger.log(self.eval(), step=0)
-
-        for epoch in range(self.cfgs.train_cfgs.epochs):
-            self.model.train()
-
-            for batch in self.train_dataloader:
-                info = self.train_step(batch)
-                torch.cuda.empty_cache()
-
-                self.global_step += 1
-                progress_bar.set_description(
-                    f'Training {epoch + 1}/{self.cfgs.train_cfgs.epochs} epoch '
-                    f'(loss {info["train/loss"]:.4f})',
-                )
-                progress_bar.update(1)
-
-                info['train/epoch'] = self.global_step / len(self.train_dataloader)
-                self.logger.log(info, step=self.global_step)
-
-                if self.global_step % self.cfgs.logger_cfgs.save_interval == 0:
-                    self.logger.print(f'Saving checkpoint at step {self.global_step} ...')
-                    self.save(tag=self.global_step)
-                    self.logger.print('Checkpoint saved.')
-
-                if (
-                    self.cfgs.data_cfgs.eval_datasets
-                    and self.cfgs.train_cfgs.eval_strategy == 'steps'
-                    and self.global_step % self.cfgs.train_cfgs.eval_interval == 0
-                ):
-                    self.logger.print(f'\n***** Evaluating at step {self.global_step} *****')
-                    self.logger.log(self.eval(), step=self.global_step)
-
-            if self.cfgs.data_cfgs.eval_datasets and self.cfgs.train_cfgs.eval_strategy == 'epoch':
-                self.logger.print(
-                    f'\n***** Evaluating at epoch {epoch + 1}/{self.cfgs.train_cfgs.epochs} *****',
-                )
-                self.logger.log(self.eval(), step=self.global_step)
-
-            self.model.tput_timer.update_epoch_count()
-
-    def save(
-        self,
-        model: deepspeed.DeepSpeedEngine | None = None,
-        tag: int | None = None,
-    ) -> None:
-        """Save model and tokenizer in Hugging Face format."""
-        dist.barrier()
-
-        if model is None:
-            model = self.model  # pylint: disable=no-member
-
-        self.logger.print(f'Saving model to "{self.cfgs.logger_cfgs.output_dir}" ...')
-
-        output_config_file = os.path.join(self.cfgs.logger_cfgs.output_dir, CONFIG_NAME)
-        model_to_save: PreTrainedModel = getattr(model, 'module', model)
-
-        if is_main_process():
-            model_to_save.config.to_json_file(output_config_file)
-            self.tokenizer.save_pretrained(self.cfgs.logger_cfgs.output_dir)
-
-        self.logger.print('Saving 16-bit model...')
-        save_file_name = f'pytorch_model_{tag}.bin' if tag else 'pytorch_model.bin'
-        model.save_16bit_model(self.cfgs.logger_cfgs.output_dir, save_filename=save_file_name)
-
-        self.logger.print('Model saved!')
-
-
-def main():
-    # setup distribution training
-    deepspeed.init_distributed()
-    current_device = get_current_device()
-    torch.cuda.set_device(current_device)
-
-    # read default configs from the yaml file
-    dict_cfgs, ds_cfgs = read_cfgs(mode='train', task='dpo')
-
-    # get custom configs from command line
-    parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter)
-    _, unparsed_args = parser.parse_known_args()
-    keys = [k[2:] for k in unparsed_args[1::2]]
-    values = list(unparsed_args[2::2])
-    unparsed_args = dict(zip(keys, values))
-    for k, v in unparsed_args.items():
-        dict_cfgs = update_dict(dict_cfgs, custom_cfgs_to_dict(k, v))
-
-    # setup training
-    cfgs = dict_to_namedtuple(dict_cfgs)
-    seed_everything(cfgs.train_cfgs.seed)
-
-    # finetune the model
-    trainer = DPOTrainer(cfgs=cfgs, ds_cfgs=ds_cfgs)
-    trainer.train()
-    trainer.save()
-
-
-if __name__ == '__main__':
-    sys.exit(main())
