+++ align_anything/trainers/text_image_to_text/__init__.py
