+++ align_anything/trainers/text_to_image/dpo_diffusion.py
@@ -12,7 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Trainer for supervised training."""
+"""Trainer for diffusion DPO."""
+
 
 import os
 import sys
@@ -21,15 +22,17 @@ from typing import Any
 import deepspeed
 import torch
 import torch.nn.functional as F
+import torch.distributed as dist
 
 from accelerate import Accelerator
 
-import torch.distributed as dist
+from tqdm import tqdm
 
-from align_anything.datasets.text_to_image import AnyToImageDataset, AnyToImageBatch
+from align_anything.datasets.text_to_image import PreferenceDataset, PreferenceBatch
 from align_anything.models.pretrained_model import load_pretrained_diffusion_models
-from align_anything.trainers.base import SupervisedTrainer
-from align_anything.utils.multi_process import get_current_device
+from align_anything.trainers.base import SupervisedTrainerBase
+from align_anything.utils.multi_process import get_current_device, get_all_reduce_mean, is_main_process
+from align_anything.utils.process_image import get_image_processor
 from align_anything.utils.tools import (
     custom_cfgs_to_dict,
     dict_to_namedtuple,
@@ -40,11 +43,9 @@ from align_anything.utils.tools import (
     prepare_accelerate_train_cfgs
 )
 
-from diffusers import StableDiffusionPipeline
-from diffusers.utils.torch_utils import is_compiled_module
 
 
-class DiffusionTrainer(SupervisedTrainer):
+class DPOTrainer(SupervisedTrainerBase):
 
     def __init__(self, cfgs) -> None:
         """Initialize the SFT trainer."""
@@ -55,6 +56,7 @@ class DiffusionTrainer(SupervisedTrainer):
             gradient_accumulation_steps=self.muti_process_cfgs['gradient_accumulation_steps'],
             mixed_precision=self.muti_process_cfgs['mixed_precision'],
         )
+        self.dtype = torch.bfloat16 if self.muti_process_cfgs['mixed_precision'] else torch.float32
         if torch.backends.mps.is_available():
             self.accelerator.native_amp = False
 
@@ -76,31 +78,48 @@ class DiffusionTrainer(SupervisedTrainer):
         """Initialize model and tokenizer."""
         self.model, self.vae, self.text_encoder, self.noise_scheduler, self.tokenizer = load_pretrained_diffusion_models(
             self.cfgs.model_cfgs.model_name_or_path,
-            model_max_length=self.cfgs.model_cfgs.model_max_length,
-            padding_side='right',
             trust_remote_code=True,
+            freeze_unet=self.cfgs.train_cfgs.freeze_unet,
+            lora_unet=self.cfgs.train_cfgs.lora_unet,
+            dtype=self.dtype
         )
+        self.processor = get_image_processor(resolution=int(self.cfgs.train_cfgs.resolution))
 
     def init_datasets(self) -> None:
         """Initialize training and evaluation datasets."""
-        self.train_dataloader, self.eval_dataloader = self.get_dataloaders(AnyToImageDataset, AnyToImageDataset)
+        self.train_dataloader, self.eval_dataloader = self.get_dataloaders(PreferenceDataset, PreferenceDataset)
 
     def init_engines(self) -> None:
-        """Initialize DeepSpeed engines."""
+        """Initialize Accelerate engines."""
         self.init_accelerate_engines()
 
-    def loss(self, sft_batch: AnyToImageBatch) -> dict[str, torch.Tensor]:
-        """Loss function for supervised finetuning."""
-        latents = self.vae.encode(sft_batch["pixel_values"].to(self.vae.dtype)).latent_dist.sample()
+    def loss(self, batch: PreferenceBatch) -> dict[str, torch.Tensor]:
+        """Loss function for DPO finetuning."""
+        pixel_values = batch["pixel_values"].to(dtype=self.dtype)
+        feed_pixel_values = torch.cat(pixel_values.chunk(2, dim=1))
+        latents = []
+        for i in range(0, feed_pixel_values.shape[0], self.cfgs.train_cfgs.vae_encode_batch_size):
+            latents.append(
+                self.vae.encode(feed_pixel_values[i : i + self.cfgs.train_cfgs.vae_encode_batch_size]).latent_dist.sample()
+            )
+        latents = torch.cat(latents, dim=0)
         latents = latents * self.vae.config.scaling_factor
 
-        noise = torch.randn_like(latents)
-        batch_size = latents.shape[0]
-        timesteps = torch.randint(0, self.noise_scheduler.config.num_train_timesteps, (batch_size,), device=latents.device)
-        timesteps = timesteps.long()
-        noisy_latents = self.noise_scheduler.add_noise(latents, noise, timesteps)
+        noise = torch.randn_like(latents).chunk(2)[0].repeat(2, 1, 1, 1)
+        batch_size = latents.shape[0] // 2
         
-        encoder_hidden_states = self.text_encoder(sft_batch["input_ids"], return_dict=False)[0]
+        timesteps = torch.randint(
+            0, 
+            self.noise_scheduler.config.num_train_timesteps, 
+            (batch_size,), 
+            device=latents.device, 
+            dtype=torch.long
+        ).repeat(2)
+
+        encoder_hidden_states = self.text_encoder(batch["input_ids"], attention_mask=None)[0].repeat(2, 1, 1)
+        noisy_latents = self.noise_scheduler.add_noise(latents, noise, timesteps)
+        model_pred = self.model(noisy_latents, timesteps, encoder_hidden_states).sample
+        
         if self.noise_scheduler.config.prediction_type == "epsilon":
             target = noise
         elif self.noise_scheduler.config.prediction_type == "v_prediction":
@@ -108,49 +127,124 @@ class DiffusionTrainer(SupervisedTrainer):
         else:
             raise ValueError(f"Unknown prediction type {self.noise_scheduler.config.prediction_type}")
         
-        model_pred = self.model(noisy_latents, timesteps, encoder_hidden_states, return_dict=False)[0]
-        loss = F.mse_loss(model_pred.float(), target.float(), reduction="mean")
+        model_losses = F.mse_loss(model_pred.float(), target.float(), reduction="none")
+        model_losses = model_losses.mean(dim=list(range(1, len(model_losses.shape))))
+        model_losses_w, model_losses_l = model_losses.chunk(2)
+        
+        # For logging
+        model_diff = model_losses_w - model_losses_l  # These are both LBS (as is t)
+
+        # Reference model predictions
+        if self.cfgs.train_cfgs.lora_unet:
+            self.accelerator.unwrap_model(self.model).disable_adapters()
+        with torch.no_grad():
+            ref_preds = self.model(
+                noisy_latents,
+                timesteps,
+                encoder_hidden_states,
+            ).sample.detach()
+            ref_loss = F.mse_loss(ref_preds.float(), target.float(), reduction="none")
+            ref_loss = ref_loss.mean(dim=list(range(1, len(ref_loss.shape))))
+
+            ref_losses_w, ref_losses_l = ref_loss.chunk(2)
+            ref_diff = ref_losses_w - ref_losses_l
+        
+        if self.cfgs.train_cfgs.lora_unet:
+            self.accelerator.unwrap_model(self.model).enable_adapters()
+        
+        logits = ref_diff - model_diff
+        if self.cfgs.train_cfgs.loss_type == "sigmoid":
+            loss = -F.logsigmoid(self.cfgs.train_cfgs.beta_coeff * logits).mean()
+        elif self.cfgs.train_cfgs.loss_type == "hinge":
+            loss = torch.relu(1 - self.cfgs.train_cfgs.beta_coeff * logits).mean()
+        else:
+            raise ValueError(f"Unknown loss type {self.cfgs.train_cfgs.loss_type}")
+
+        implicit_acc = (logits > 0).sum().float() / logits.size(0)
+        implicit_acc += 0.5 * (logits == 0).sum().float() / logits.size(0)
         
         return {
             'loss': loss,
+            'reward_accuracy': implicit_acc,
         }
 
-    def train_step(self, sft_batch: AnyToImageBatch) -> dict[str, Any]:
+    def train_step(self, batch: PreferenceBatch) -> dict[str, Any]:
         """Performs a single training step."""
-        loss = self.loss(sft_batch)['loss']
+        loss = self.loss(batch)['loss']
+        reward_accuracy = self.loss(batch)['reward_accuracy']
+        
         self.accelerator.backward(loss)
         if self.accelerator.sync_gradients:
-            self.accelerator.clip_grad_norm_(self.model.parameters(), self.cfgs.train_cfgs.max_grad_norm)
+            self.accelerator.clip_grad_norm_(self.params_to_optimize, self.cfgs.train_cfgs.max_grad_norm)
         self.optimizer.step()
         self.lr_scheduler.step()
         self.optimizer.zero_grad()
+        
+        with torch.no_grad():
+            loss = get_all_reduce_mean(loss)
+            reward_accuracy = get_all_reduce_mean(reward_accuracy)
+        
         return {
             'train/loss': loss.item(),
             'train/lr': self.optimizer.param_groups[0]['lr'],
+            'train/reward_accuracy': reward_accuracy.item(),
         }
 
-    def save(
-        self,
-        model: deepspeed.DeepSpeedEngine | None = None,
-        tag: int | None = None,
-    ) -> None:
-        """Save model and tokenizer in Hugging Face format."""
-        self.accelerator.wait_for_everyone()
-        save_dir = os.path.join(self.cfgs.logger_cfgs.output_dir, f'epoch_{tag}')
-        if self.accelerator.is_main_process:
-            model = self.accelerator.unwrap_model(self.model)
-            model = model._orig_mod if is_compiled_module(model) else model
-            pipeline = StableDiffusionPipeline.from_pretrained(
-                self.cfgs.model_cfgs.model_name_or_path,
-                text_encoder=self.text_encoder,
-                vae=self.vae,
-                unet=model,
-                revision=self.cfgs.train_cfgs.revision,
-                variant=self.cfgs.train_cfgs.variant,
-            )
-            pipeline.save_pretrained(save_dir)
+    def train(self) -> None:
+        """Train the model."""
+        self.logger.print('***** Running training *****')
 
-        self.logger.print('Model saved!')
+        progress_bar = tqdm(
+            total=self.cfgs.train_cfgs.epochs * len(self.train_dataloader),
+            desc=f'Training 1/{self.cfgs.train_cfgs.epochs} epoch',
+            position=0,
+            leave=True,
+            disable=not is_main_process(),
+        )
+
+        if self.cfgs.data_cfgs.eval_datasets:
+            self.logger.print('\n***** Evaluating at the beginning *****')
+            self.logger.log(self.eval(), step=0)
+
+        for epoch in range(int(self.cfgs.train_cfgs.epochs)):
+            self.model.train()
+
+            for batch in self.train_dataloader:
+                info = self.train_step(batch)
+                torch.cuda.empty_cache()
+
+                self.global_step += 1
+                progress_bar.set_description(
+                    f'Training {epoch + 1}/{self.cfgs.train_cfgs.epochs} epoch '
+                    f'(loss {info["train/loss"]:.4f})',
+                )
+                progress_bar.update(1)
+
+                info['train/epoch'] = self.global_step / len(self.train_dataloader)
+                self.logger.log(info, step=self.global_step)
+
+                if self.global_step % self.cfgs.logger_cfgs.save_interval == 0:
+                    self.logger.print(f'Saving checkpoint at step {self.global_step} ...')
+                    self.save(tag=self.global_step)
+                    self.logger.print('Checkpoint saved.')
+
+                if (
+                    self.cfgs.data_cfgs.eval_datasets
+                    and self.cfgs.train_cfgs.eval_strategy == 'steps'
+                    and self.global_step % self.cfgs.train_cfgs.eval_interval == 0
+                ):
+                    self.logger.print(f'\n***** Evaluating at step {self.global_step} *****')
+                    self.logger.log(self.eval(), step=self.global_step)
+
+            if self.cfgs.data_cfgs.eval_datasets and self.cfgs.train_cfgs.eval_strategy == 'epoch':
+                self.logger.print(
+                    f'\n***** Evaluating at epoch {epoch + 1}/{self.cfgs.train_cfgs.epochs} *****',
+                )
+                self.logger.log(self.eval(), step=self.global_step)
+
+    def save(self, tag: int | None = None) -> None:
+        """Save the stable diffusion pipeline in Hugging Face format."""
+        self.save_diffusers(tag=tag)
 
 
 def main():
@@ -159,7 +253,8 @@ def main():
     torch.cuda.set_device(current_device)
 
     # read default configs from the yaml file
-    dict_cfgs, _ = read_cfgs(mode='train', task='sft')
+    task = os.path.join('text_to_image', 'dpo')
+    dict_cfgs, _ = read_cfgs(mode='train', task=task)
     unparsed_args = parse_unknown_args()
     for k, v in unparsed_args.items():
         dict_cfgs = update_dict(dict_cfgs, custom_cfgs_to_dict(k, v))
@@ -169,7 +264,7 @@ def main():
     seed_everything(cfgs.train_cfgs.seed)
 
     # finetune the model
-    trainer = DiffusionTrainer(cfgs=cfgs)
+    trainer = DPOTrainer(cfgs=cfgs)
     trainer.train()
     trainer.save()
     trainer.accelerator.end_training()
