+++ align_anything/trainers/text_to_text/dpo.py
@@ -18,23 +18,19 @@
 import argparse
 import os
 import sys
-from datetime import datetime
 from typing import Any
 
 import deepspeed
 import torch
 import torch.distributed as dist
 import torch.nn.functional as F
-from deepspeed.ops.adam import FusedAdam
-from torch.utils.data import DataLoader
-from torch.utils.data.distributed import DistributedSampler
 from tqdm import tqdm
-from transformers import CONFIG_NAME, AutoModelForCausalLM, PreTrainedModel, get_scheduler
+from transformers import AutoModelForCausalLM
 from transformers.integrations.deepspeed import HfDeepSpeedConfig
 
 from align_anything.datasets.text_to_text.preference import PreferenceBatch, PreferenceDataset
 from align_anything.models.pretrained_model import load_pretrained_models
-from align_anything.utils.logger import Logger
+from align_anything.trainers.base import SupervisedTrainerBase
 from align_anything.utils.multi_process import (
     get_all_reduce_mean,
     get_current_device,
@@ -44,8 +40,6 @@ from align_anything.utils.tools import (
     custom_cfgs_to_dict,
     dict_to_namedtuple,
     gather_log_probabilities,
-    get_optimizer_grouped_parameters,
-    namedtuple_to_dict,
     prepare_ds_eval_cfgs,
     prepare_ds_train_cfgs,
     read_cfgs,
@@ -54,7 +48,7 @@ from align_anything.utils.tools import (
 )
 
 
-class DPOTrainer:
+class DPOTrainer(SupervisedTrainerBase):
 
     def __init__(self, cfgs, ds_cfgs) -> None:
         """Initialize trainer."""
@@ -77,19 +71,6 @@ class DPOTrainer:
         """Initial configuration checking."""
         return
 
-    def init_logger(self) -> None:
-        """Set logger."""
-        logger_cfgs = self.cfgs.logger_cfgs
-        time = datetime.now().strftime('%Y-%m-%d-%H-%M-%S')
-
-        self.logger = Logger(
-            log_type=logger_cfgs.log_type,
-            log_dir=logger_cfgs.output_dir,
-            log_project=logger_cfgs.log_project,
-            log_run_name=f'{logger_cfgs.log_run_name}-{self.cfgs.data_cfgs.train_datasets}-{time}',
-            config=namedtuple_to_dict(self.cfgs),
-        )
-
     def init_models(self) -> None:
         """Initialize model and tokenizer."""
         if self.ds_train_cfgs['zero_optimization']['stage'] == 3:
@@ -101,8 +82,6 @@ class DPOTrainer:
             model_max_length=self.cfgs.model_cfgs.model_max_length,
             padding_side='left',
             trust_remote_code=self.cfgs.train_cfgs.trust_remote_code,
-            freeze_mm_proj=self.cfgs.train_cfgs.freeze_mm_proj,
-            freeze_vision_tower=self.cfgs.train_cfgs.freeze_vision_tower,
         )
         self.reference_model, _, _ = load_pretrained_models(
             self.cfgs.model_cfgs.model_name_or_path,
@@ -113,81 +92,15 @@ class DPOTrainer:
 
     def init_datasets(self) -> None:
         """Initialize training and evaluation datasets."""
-        train_dataset = PreferenceDataset(
-            path=self.cfgs.data_cfgs.train_datasets,
-            template=self.cfgs.data_cfgs.train_template,
-            tokenizer=self.tokenizer,
-            processor=self.processor,
-            size=self.cfgs.data_cfgs.train_size,
-            split=self.cfgs.data_cfgs.train_split,
-            subset=self.cfgs.data_cfgs.train_subset,
-            data_files=self.cfgs.data_cfgs.train_data_files,
-        )
-        self.train_dataloader = DataLoader(
-            train_dataset,
-            collate_fn=train_dataset.get_collator(),
-            sampler=DistributedSampler(train_dataset, shuffle=True),
-            batch_size=self.cfgs.train_cfgs.per_device_train_batch_size,
-        )
-        if self.cfgs.data_cfgs.eval_datasets:
-            eval_dataset = PreferenceDataset(
-                path=self.cfgs.data_cfgs.eval_datasets,
-                template=self.cfgs.data_cfgs.eval_template,
-                tokenizer=self.tokenizer,
-                processor=self.processor,
-                size=self.cfgs.data_cfgs.eval_size,
-                split=self.cfgs.data_cfgs.eval_split,
-                subset=self.cfgs.data_cfgs.eval_subset,
-                data_files=self.cfgs.data_cfgs.eval_data_files,
-            )
-            self.eval_dataloader = DataLoader(
-                eval_dataset,
-                collate_fn=eval_dataset.get_collator(),
-                sampler=DistributedSampler(eval_dataset, shuffle=True),
-                batch_size=self.cfgs.train_cfgs.per_device_train_batch_size,
-            )
-        self.split_token = train_dataset.template.split_token
+        self.train_dataloader, self.eval_dataloader = self.get_dataloaders(PreferenceDataset, PreferenceDataset)
 
     def init_engines(self) -> None:
         """Initialize DeepSpeed engines."""
-        num_update_steps_per_epoch = (
-            len(self.train_dataloader) + self.cfgs.train_cfgs.gradient_accumulation_steps - 1
-        ) // self.cfgs.train_cfgs.gradient_accumulation_steps
-        total_training_steps = self.cfgs.train_cfgs.epochs * num_update_steps_per_epoch
-
-        optimizer_grouped_parameters = get_optimizer_grouped_parameters(
-            self.model,
-            self.cfgs.train_cfgs.weight_decay,
-        )
-        optimizer = FusedAdam(
-            optimizer_grouped_parameters,
-            lr=self.cfgs.train_cfgs.learning_rate,
-            betas=self.cfgs.train_cfgs.adam_betas,
-        )
-
-        num_warmup_steps = int(self.cfgs.train_cfgs.lr_warmup_ratio * total_training_steps)
-        lr_scheduler = get_scheduler(
-            name=self.cfgs.train_cfgs.lr_scheduler_type,
-            optimizer=optimizer,
-            num_warmup_steps=num_warmup_steps,
-            num_training_steps=total_training_steps,
-        )
-
-        self.model, *_ = deepspeed.initialize(
-            model=self.model,
-            optimizer=optimizer,
-            config=self.ds_train_cfgs,
-            lr_scheduler=lr_scheduler,
-            dist_init_required=True,
-        )
-
+        self.init_deepspeed_engines()
         self.reference_model, *_ = deepspeed.initialize(
             model=self.reference_model,
             config=self.ds_eval_cfgs,
         )
-
-        if self.cfgs.train_cfgs.gradient_checkpointing:
-            self.model.gradient_checkpointing_enable()
 
     @staticmethod
     def compute_log_probs(
@@ -244,7 +157,6 @@ class DPOTrainer:
             better_seq_slice = slice(diverge_index, better_end_index + 1)
             worse_seq_slice = slice(diverge_index, worse_end_index + 1)
 
-            # size = ()
             better_log_prob = better_sequence_log_probs[i, better_seq_slice].sum(dim=-1)
             worse_log_prob = worse_sequence_log_probs[i, worse_seq_slice].sum(dim=-1)
             ref_better_log_prob = ref_better_sequence_log_probs[i, better_seq_slice].sum(dim=-1)
@@ -261,7 +173,6 @@ class DPOTrainer:
                 self.cfgs.train_cfgs.scale_coeff * better_log_ratio.detach(),
             )
             worse_sample_rewards.append(self.cfgs.train_cfgs.scale_coeff * worse_log_ratio.detach())
-
         loss = torch.stack(losses).mean()  # size = ()
         better_sample_reward = torch.stack(better_sample_rewards)  # size = (B,)
         worse_sample_reward = torch.stack(worse_sample_rewards)  # size = (B,)
@@ -312,11 +223,6 @@ class DPOTrainer:
             'train/lr': self.model.optimizer.param_groups[0]['lr'],
         }
 
-    @torch.no_grad()
-    def eval(self) -> dict[str, Any]:
-        """Evaluate the model on the evaluation dataset."""
-        return {}
-
     def train(self) -> None:
         """Train the model."""
         self.logger.print('***** Running training *****')
@@ -333,7 +239,7 @@ class DPOTrainer:
             self.logger.print('\n***** Evaluating at the beginning *****')
             self.logger.log(self.eval(), step=0)
 
-        for epoch in range(self.cfgs.train_cfgs.epochs):
+        for epoch in range(int(self.cfgs.train_cfgs.epochs)):
             self.model.train()
 
             for batch in self.train_dataloader:
@@ -368,8 +274,12 @@ class DPOTrainer:
                     f'\n***** Evaluating at epoch {epoch + 1}/{self.cfgs.train_cfgs.epochs} *****',
                 )
                 self.logger.log(self.eval(), step=self.global_step)
-
             self.model.tput_timer.update_epoch_count()
+
+    @torch.no_grad()
+    def eval(self) -> dict[str, Any]:
+        """Evaluate the model on the evaluation dataset."""
+        return {}
 
     def save(
         self,
@@ -377,25 +287,7 @@ class DPOTrainer:
         tag: int | None = None,
     ) -> None:
         """Save model and tokenizer in Hugging Face format."""
-        dist.barrier()
-
-        if model is None:
-            model = self.model  # pylint: disable=no-member
-
-        self.logger.print(f'Saving model to "{self.cfgs.logger_cfgs.output_dir}" ...')
-
-        output_config_file = os.path.join(self.cfgs.logger_cfgs.output_dir, CONFIG_NAME)
-        model_to_save: PreTrainedModel = getattr(model, 'module', model)
-
-        if is_main_process():
-            model_to_save.config.to_json_file(output_config_file)
-            self.tokenizer.save_pretrained(self.cfgs.logger_cfgs.output_dir)
-
-        self.logger.print('Saving 16-bit model...')
-        save_file_name = f'pytorch_model_{tag}.bin' if tag else 'pytorch_model.bin'
-        model.save_16bit_model(self.cfgs.logger_cfgs.output_dir, save_filename=save_file_name)
-
-        self.logger.print('Model saved!')
+        self.save_transformers(model=model, tag=tag)
 
 
 def main():
@@ -405,7 +297,8 @@ def main():
     torch.cuda.set_device(current_device)
 
     # read default configs from the yaml file
-    dict_cfgs, ds_cfgs = read_cfgs(mode='train', task='dpo')
+    task = os.path.join('text_to_text', 'dpo')
+    dict_cfgs, ds_cfgs = read_cfgs(mode='train', task=task)
 
     # get custom configs from command line
     parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter)
