+++ align_anything/trainers/text_to_text/simpo.py
@@ -213,8 +213,6 @@ class SimPOTrainer:
             worse_sequence_log_probs,  # size = (B, L - 1)
         ) = sequence_log_probs.chunk(chunks=2, dim=0)
 
-        
-
         losses = []
         better_sample_rewards = []
         worse_sample_rewards = []
@@ -397,7 +395,8 @@ def main():
     torch.cuda.set_device(current_device)
 
     # read default configs from the yaml file
-    dict_cfgs, ds_cfgs = read_cfgs(mode='train', task='simpo')
+    task = os.path.join('text_to_text', 'simpo')
+    dict_cfgs, ds_cfgs = read_cfgs(mode='train', task=task)
 
     # get custom configs from command line
     parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter)
